ibqueryerrors.c: Add support for additional counters in PortCountersExtended

+++ src/ibdiag_common.c
@@ -439,7 +439,7 @@ conv_cnt_human_readable(uint64_t val64, float *val, int data)
 {
 	uint64_t tmp = val64;
 	int ui = 0;
-	int div = 1;
+	uint64_t div = 1;
 
 	tmp /= 1024;
 	while (tmp) {
